+++ src/Entity/Track.php
@@ -7,6 +7,7 @@ use App\Entity\Track\OptimizedPoint;
 use App\Entity\Track\Version;
 use App\Entity\User\User;
 use App\Entity\Video\Youtube;
+use App\EntityTraits\NameableTrait;
 use DateTime;
 use Doctrine\Common\Collections\ArrayCollection;
 use Doctrine\ORM\Mapping as ORM;
@@ -16,6 +17,8 @@ use Doctrine\ORM\Mapping as ORM;
  */
 class Track
 {
+    use NameableTrait;
+
     public const TYPE_CYCLING = 1;
     public const TYPE_HIKING = 2;
 
@@ -38,16 +41,6 @@ class Track
      * @ORM\Column(type="string", nullable=true)
      */
     private $slug;
-
-    /**
-     * @ORM\Column(type="string", nullable=true)
-     */
-    private $nameEn;
-
-    /**
-     * @ORM\Column(type="string", nullable=true)
-     */
-    private $nameBg;
 
     /**
      * @ORM\Column(type="boolean", nullable=false)
@@ -148,42 +141,6 @@ class Track
         return $this->sendBy;
     }
 
-    /**
-     * Get localized name.
-     * If localized name doesn't exists - fallback to EN
-     */
-    public function getName(string $locale): ?string
-    {
-        if ($locale === 'bg' && $this->nameBg !== null) {
-            return $this->nameBg;
-        }
-
-        if ($this->nameEn === null) {
-            return $this->nameBg;
-        }
-
-        return $this->nameEn;
-    }
-
-    public function getNameEn(): ?string
-    {
-        return $this->nameEn;
-    }
-
-    public function getNameBg(): ?string
-    {
-        return $this->nameBg;
-    }
-
-    public function setNameEn(string $nameEn = null): void
-    {
-        $this->nameEn = $nameEn;
-    }
-
-    public function setNameBg(string $nameBg = null): void
-    {
-        $this->nameBg = $nameBg;
-    }
 
     public function getId(): ?string
     {
