+++ src/Form/Type/Place.php
@@ -4,6 +4,7 @@ namespace App\Form\Type;
 
 use Symfony\Component\Form\AbstractType;
 use Symfony\Component\Form\Extension\Core\Type\NumberType;
+use Symfony\Component\Form\Extension\Core\Type\TextType;
 use Symfony\Component\Form\FormBuilderInterface;
 use Symfony\Component\Validator\Constraints\Range;
 
@@ -11,6 +12,24 @@ class Place extends AbstractType
 {
     public function buildForm(FormBuilderInterface $builder, array $options)
     {
+        $builder->add(
+            'nameEn',
+            TextType::class,
+            [
+                'required' => false,
+                'label' => 'Name in English',
+            ]
+        );
+
+        $builder->add(
+            'nameBg',
+            TextType::class,
+            [
+                'required' => false,
+                'label' => 'Name in Bulgarian',
+            ]
+        );
+
         $coordinatesConstraints = [
             new Range([
                 'min' => -180,
@@ -37,5 +56,7 @@ class Place extends AbstractType
                 'constraints' => $coordinatesConstraints,
             ]
         );
+
+
     }
 }
