+++ templates/place/edit.html.twig
@@ -88,7 +88,7 @@
                 }
             };
 
-            new syncInputAndMap(popup, document.getElementById('lat'), document.getElementById('lng'));
+            new syncInputAndMap(popup, document.getElementById('place_lat'), document.getElementById('place_lng'));
         })();
     </script>
 {% endblock %}
